// examples/games/lootsurvivor/src/app/types/events.ts
type u128 = number;
type u64 = number;
type u32 = number;
type u16 = number;
type u8 = number;

type AdventurerMetadata = {
  birthDate: string;
  deathDate: string;
  adventurerEntropy: u64;
  itemSpecialsSeed: u16;
  rankAtDeath: u8;
  delayStatReveal: boolean;
  goldenTokenId: u8;
  launchTournamentWinnerTokenId: u32;
};

type Stats = {
  strength: u8;
  dexterity: u8;
  vitality: u8;
  intelligence: u8;
  wisdom: u8;
  charisma: u8;
  luck: u8;
};

type Equipment = {
  weapon: LootStatistics;
  chest: LootStatistics;
  head: LootStatistics;
  waist: LootStatistics;
  foot: LootStatistics;
  hand: LootStatistics;
  neck: LootStatistics;
  ring: LootStatistics;
};

type LootStatistics = {
  id: u8;
  xp: u16;
};

type Adventurer = {
  health: u16;
  xp: u16;
  gold: u16;
  beastHealth: u16;
  statUpgradesAvailable: u8;
  stats: Stats;
  equipment: Equipment;
  battleActionCount: u8;
  mutated: boolean;
  awaitingItemSpecials: boolean;
};
// ...
type ContractAddress = string;

export type AdventurerState = {
  owner: ContractAddress;
  adventurerId: u128;
  adventurerEntropy: string;
  adventurer: Adventurer;
};